新增sidebar component 讓他共用於其他網頁
